Mengaplikasikan cache service ke dalam service utama (Notes and Collaborations)

+++ src/services/postgres/NotesService.js
@@ -6,7 +6,7 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class NotesService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._pool = new Pool({
       user: process.env.PGUSER,
       host: process.env.PGHOST,
@@ -15,6 +15,7 @@ class NotesService {
       port: process.env.PGPORT,
     });
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
   }
 
   async addNote({title, body, tags, owner}) {
@@ -33,19 +34,31 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan.');
     }
 
+    await this._cacheService.del(`notes:${owner}`);
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM
-      notes LEFT JOIN collaborations ON collaborations.note_id = notes.id
-      WHERE notes.owner=$1 OR collaborations.user_id=$1
-      GROUP BY notes.id`,
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
+    try {
+      const resultNoteFromCache = await this._cacheService.get(`notes:${owner}`);
+      return JSON.parse(resultNoteFromCache);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM
+        notes LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner=$1 OR collaborations.user_id=$1
+        GROUP BY notes.id`,
+        values: [owner],
+      };
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      await this._cacheService.set({
+        key: `notes:${owner}`,
+        value: JSON.stringify(mappedResult),
+      });
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -68,7 +81,7 @@ class NotesService {
   async editNoteById(id, {title, body, tags}) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id',
+      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
@@ -77,11 +90,13 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
+
+    await this._cacheService.del(`notes:${result.rows[0].owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id=$1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id=$1 RETURNING id, owner',
       values: [id],
     };
 
@@ -90,6 +105,8 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+
+    await this._cacheService.del(`notes:${result.rows[0].owner}`);
   }
 
   async verifyNoteOwner(id, owner) {
